[login-Refactor] 로그인 리팩토링 -> header의 버튼에서 이루어진 로그인을 분리함 - <이전> header에 로그인 버튼을 클릭하면 google로그인을 바로 하게 설정 - <리팩토링> 로그인 버튼 클릭시 로그인 방법을 선택할 수 있는 모달이 뜨게 함 -> 유저가 로그인을 선택하면 이후 로그인이 되고, 로그아웃 버튼이 헤더에 반영됨 useAuth컨텍스트에 있는 authUser(google로그인 여부)를 확인해 로그인/로그아웃 버튼을 토글할 수 있게 함함
