+++ src/shared/Layout/styles/menuModal.module.css
@@ -37,3 +37,21 @@
   font-size: 1rem;
   text-decoration: none;
 }
+
+/* login */
+.loginButtons {
+  margin-top: 2rem;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+.loginButtons > button {
+  width: 100%;
+  height: 3rem;
+  border: 1px solid var(--color-primary);
+  border-radius: 0.5rem;
+}
+.loginButtons > button:hover {
+  background-color: var(--color-primary-light);
+  cursor: pointer;
+}
